requirements.txt
asgiref==3.2.10
beautifulsoup4==4.8.2
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.8
django-allauth==0.42.0
django-ckeditor==5.9.0
django-cors-headers==3.4.0
django-cuser==2017.3.16
django-heroku==0.3.1
django-js-asset==1.2.2
django-modelcluster==5.0.2
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-taggit==1.3.0
django-tinymce==3.0.2
django-treebeard==4.3.1
django-wysiwyg==0.8.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
draftjs-exporter==2.1.7
gunicorn==20.0.4
html5lib==1.1
idna==2.10
l18n==2018.5
modeltranslation==0.25
mysql==0.0.2
mysql-connector-python==8.0.21
mysqlclient==2.0.1
oauthlib==3.1.0
Pillow==7.2.0
protobuf==3.12.2
psycopg2==2.8.5
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
Unidecode==1.1.1
urllib3==1.25.9
webencodings==0.5.1
whitenoise==5.2.0
Willow==1.3
XlsxWriter==1.2.9
